chore: :recycle: Add examples for keras-tuner-extensionpack
